+++ shop/cascade/catalog.py
@@ -8,7 +8,9 @@ from django.template.loader import select_template
 from django.utils.translation import ugettext_lazy as _
 from cms.plugin_pool import plugin_pool
 from cms.utils.compat.dj import is_installed
+from cmsplugin_cascade.mixins import WithSortableInlineElementsMixin
 from cmsplugin_cascade.models import SortableInlineCascadeElement
+
 from shop import app_settings
 from shop.models.product import ProductModel
 from .plugin_base import ShopPluginBase, ProductSelectField
@@ -93,7 +95,7 @@ class ProductGalleryInline(SortableInlineAdminMixin, StackedInline):
     verbose_name_plural = _("Product Gallery")
 
 
-class ShopProductGallery(ShopPluginBase):
+class ShopProductGallery(WithSortableInlineElementsMixin, ShopPluginBase):
     name = _("Product Gallery")
     require_parent = True
     parent_classes = ('BootstrapColumnPlugin',)
@@ -111,8 +113,6 @@ class ShopProductGallery(ShopPluginBase):
         ])
 
     def render(self, context, instance, placeholder):
-        from shop.rest.bases import get_product_summary_serializer_class
-
         product_ids = []
         for instance in instance.sortinline_elements.all():
             try:
@@ -120,7 +120,7 @@ class ShopProductGallery(ShopPluginBase):
             except KeyError:
                 pass
         queryset = ProductModel.objects.filter(pk__in=product_ids)
-        serializer_class = get_product_summary_serializer_class()
+        serializer_class = app_settings.PRODUCT_SUMMARY_SERIALIZER
         serialized = serializer_class(queryset, many=True, context={'request': context['request']})
         context['products'] = serialized.data
         return context