+++ shop/models/customer.py
@@ -78,13 +78,13 @@ class CustomerQuerySet(models.QuerySet):
             if field_name == 'pk':
                 field_name = opts.pk.name
             try:
-                opts.get_field_by_name(field_name)
+                opts.get_field(field_name)
                 if isinstance(lookup, get_user_model()):
                     lookup.pk  # force lazy object to resolve
                 lookup_kwargs[key] = lookup
             except FieldDoesNotExist as fdne:
                 try:
-                    get_user_model()._meta.get_field_by_name(field_name)
+                    get_user_model()._meta.get_field(field_name)
                     lookup_kwargs['user__' + key] = lookup
                 except FieldDoesNotExist:
                     raise fdne
@@ -95,6 +95,11 @@ class CustomerQuerySet(models.QuerySet):
 
 
 class CustomerManager(models.Manager):
+    """
+    Manager for the Customer database model. This manager can also cope with customers, which have
+    an entity in the database but otherwise are considered as anonymous. The username of these
+    so called unrecognized customers is a compact version of the session key.
+    """
     BASE64_ALPHABET = string.digits + string.ascii_uppercase + string.ascii_lowercase + '.@'
     REVERSE_ALPHABET = dict((c, i) for i, c in enumerate(BASE64_ALPHABET))
     BASE36_ALPHABET = string.digits + string.ascii_lowercase
@@ -135,7 +140,7 @@ class CustomerManager(models.Manager):
     def get_queryset(self):
         """
         Whenever we fetch from the Customer table, inner join with the User table to reduce the
-        number of queries to the database.
+        number of presumed future queries to the database.
         """
         qs = self._queryset_class(self.model, using=self._db).select_related('user')
         return qs
@@ -148,7 +153,7 @@ class CustomerManager(models.Manager):
 
     def _get_visiting_user(self, session_key):
         """
-        Since the Customer has a 1:1 relation with the User object, look for an entity for a
+        Since the Customer has a 1:1 relation with the User object, look for an entity of a
         User object. As its ``username`` (which must be unique), use the given session key.
         """
         username = self.encode_session_key(session_key)