+++ shop/static/shop/js/auth.js
@@ -30,6 +30,7 @@ djangoShopModule.directive('shopAuthForm', ['$window', '$http', '$timeout',
 
 			// Submit auth form data. Use `delay` in milliseconds to postpone final action.
 			scope.submitForm = function(submitURL, delay) {
+				submitURL = submitURL || attrs.formSubmitUrl;
 				$http.post(submitURL, scope.form_data).success(function(response) {
 					if (response.success) {
 						scope.success_message = response.success;
@@ -86,6 +87,16 @@ djangoShopModule.directive('shopAuthForm', ['$window', '$http', '$timeout',
 			scope.$on('$destroy', function(event) {
 				if (timer) {
 					$timeout.cancel(timer);
+				}
+			});
+
+			element.bind('keyup', function(event) {
+				if (event.which === 13) {
+					// trigger on pressed Enter key
+					scope.$apply(function() {
+						scope.$eval(scope.submitForm());
+					});
+					event.preventDefault();
 				}
 			});
 		}